refactor(grid): import presenter from index.ts #30
- do not import from presenter.ts
