+++ renderer/src/templates/grid/index.ts
@@ -1,5 +1,7 @@
 
 export { StyledComponent as Grid } from './Grid'
 export { Container as GridContainer } from './Grid'
+export * as GridPresenter from './presenter'
+export type { GridCell } from './presenter'
 export type { StyledProps as GridProps } from './Grid'
 export type { ContainerProps as GridContainerProps } from './Grid'
